+++ components/chakra/react/src/stories/code.stories.tsx
@@ -1,9 +1,10 @@
 import { Code } from '@chakra-ui/react'
 import type { Meta } from '@storybook/react'
 
-const meta: Meta<typeof Code> = {
-  component: Code,
+const meta: Meta = {
+  title: 'Components/Code',
 }
+
 export default meta
 
 export const Default = () => <Code>Code</Code>