+++ components/chakra/react/src/stories/kbd.stories.tsx
@@ -1,9 +1,10 @@
 import { Kbd } from '@chakra-ui/react'
 import type { Meta } from '@storybook/react'
 
-const meta: Meta<typeof Kbd> = {
-  component: Kbd,
+const meta: Meta = {
+  title: 'Components/Kbd',
 }
+
 export default meta
 
 export const Default = () => <Kbd>⌘ + T</Kbd>