+++ components/tailwind/react/src/stories/icon.stories.tsx
@@ -1,11 +1,10 @@
 import type { Meta } from '@storybook/react'
 import { DiamondIcon } from 'lucide-react'
 import { Stack } from 'styled-system/jsx'
-import { Icon, type IconProps } from '~/components/ui/icon'
+import { Icon } from '~/components/ui/icon'
 
-const meta: Meta<IconProps> = {
+const meta: Meta = {
   title: 'Components/Icon',
-  component: Icon,
 }
 
 export default meta