+++ components/tailwind/solid/src/stories/number-input.stories.tsx
@@ -1,9 +1,8 @@
 import type { Meta } from 'storybook-solidjs'
-import { NumberInput, type NumberInputProps } from '~/components/ui/number-input'
+import { NumberInput } from '~/components/ui/number-input'
 
-const meta: Meta<NumberInputProps> = {
+const meta: Meta = {
   title: 'Components/NumberInput',
-  component: NumberInput,
 }
 
 export default meta